prometheus_yaml/requirements.txt
python-dotenv==0.19.0
pydantic==1.8.2
fastapi==0.68.1
uvicorn==0.15.0
pyyaml
